Backend/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
defusedxml==0.7.1
Django==4.2.23
django-cors-headers==4.7.0
django-extensions==4.1
django-mongoengine==0.5.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.0
dnspython==2.7.0
docopt==0.6.2
fonttools==4.58.5
gunicorn==23.0.0
idna==3.10
mongoengine==0.29.1
num2words==0.5.14
oauthlib==3.3.1
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pymongo==4.13.2
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pywin32-ctypes==0.2.3
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
social-auth-app-django==5.0.0
social-auth-core==4.7.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
zopfli==0.2.3.post1
